Move fuse utils from commons to utils
